Initial program of the Acrobat reader connector
Summary:
This program will only work under the active window of Adobe Reader. It allows you to scroll down and up by waving in and out. Also, it will give Myo control over the mouse. Holding fist will hold a mouse click, and spreading fingers is like pressing Ctrl+Z
Quite useful if you need to read and highlight texts often. 
I hope this is helpful for some of you and the basis for improving it :D

For using:
- Download
- Open Application Manager of Myo armband
- Click "+Add"
- Select downloaded file